+++ test/switchboard/native/OptimismSwitchboardL2L1.t.sol
@@ -126,7 +126,7 @@ contract OptimismSwitchboardL2L1Test is Setup {
             switchBoardAddress_,
             DEFAULT_BATCH_LENGTH,
             uint32(remoteChainSlug_),
-            uint32(capacitorType_)
+            capacitorType_
         );
 
         scc_.siblingChainSlug = remoteChainSlug_;
